[Ingest][Endpoint] Add Ingest rest api response types for use in Endpoint (#63373)
* Added REST Response types to datasouces
* Adjust Types in Endpoint to reference those from Ingest
* Remove un-used endpoint type
