+++ x-pack/plugins/endpoint/public/applications/endpoint/store/policy_list/middleware.ts
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { MiddlewareFactory, PolicyListState, GetDatasourcesResponse } from '../../types';
+import { GetPolicyListResponse, MiddlewareFactory, PolicyListState } from '../../types';
 import { sendGetEndpointSpecificDatasources } from './services/ingest';
 import { isOnPolicyListPage, urlSearchParams } from './selectors';
 
@@ -18,7 +18,7 @@ export const policyListMiddlewareFactory: MiddlewareFactory<PolicyListState> = c
 
     if (action.type === 'userChangedUrl' && isOnPolicyListPage(state)) {
       const { page_index: pageIndex, page_size: pageSize } = urlSearchParams(state);
-      let response: GetDatasourcesResponse;
+      let response: GetPolicyListResponse;
 
       try {
         response = await sendGetEndpointSpecificDatasources(http, {
