+++ x-pack/plugins/endpoint/public/applications/endpoint/types.ts
@@ -17,11 +17,14 @@ import {
 import { EndpointPluginStartDependencies } from '../../plugin';
 import { AppAction } from './store/action';
 import { CoreStart } from '../../../../../../src/core/public';
-import { Datasource, NewDatasource } from '../../../../ingest_manager/common/types/models';
 import {
+  Datasource,
+  NewDatasource,
   GetAgentStatusResponse,
-  CreateDatasourceResponse,
-} from '../../../../ingest_manager/common/types/rest_spec';
+  GetDatasourcesResponse,
+  GetOneDatasourceResponse,
+  UpdateDatasourceResponse,
+} from '../../../../ingest_manager/common';
 
 export { AppAction };
 export type MiddlewareFactory<S = GlobalState> = (
@@ -331,24 +334,14 @@ export interface AlertingIndexUIQueryParams {
   filters?: string;
 }
 
-export interface GetDatasourcesResponse {
+export interface GetPolicyListResponse extends GetDatasourcesResponse {
   items: PolicyData[];
-  total: number;
-  page: number;
-  perPage: number;
-  success: boolean;
 }
 
-export interface GetDatasourceResponse {
+export interface GetPolicyResponse extends GetOneDatasourceResponse {
   item: PolicyData;
-  success: boolean;
 }
 
-export type UpdateDatasourceResponse = CreateDatasourceResponse & {
+export interface UpdatePolicyResponse extends UpdateDatasourceResponse {
   item: PolicyData;
-};
-
-/**
- * The PageId type is used for the payload when firing userNavigatedToPage actions
- */
-export type PageId = 'alertsPage' | 'managementPage' | 'policyListPage';
+}
