+++ x-pack/plugins/ingest_manager/common/types/rest_spec/datasource.ts
@@ -10,10 +10,23 @@ export interface GetDatasourcesRequest {
   query: ListWithKuery;
 }
 
+export interface GetDatasourcesResponse {
+  items: Datasource[];
+  total: number;
+  page: number;
+  perPage: number;
+  success: boolean;
+}
+
 export interface GetOneDatasourceRequest {
   params: {
     datasourceId: string;
   };
+}
+
+export interface GetOneDatasourceResponse {
+  item: Datasource;
+  success: boolean;
 }
 
 export interface CreateDatasourceRequest {
@@ -28,6 +41,8 @@ export interface CreateDatasourceResponse {
 export type UpdateDatasourceRequest = GetOneDatasourceRequest & {
   body: NewDatasource;
 };
+
+export type UpdateDatasourceResponse = CreateDatasourceResponse;
 
 export interface DeleteDatasourcesRequest {
   body: {
